Improved AI spawning direction picker
The spawner's direction picker now uses 'smart' methods rather than pure
chance.

This fixes the balance of spawning distance; for example, if the user is
at a dead end of a hallway, Render would previously be more likely to
spawn nearby due to the close proximity of the walls around the player.

With this change, Render is much more likely to spawn around corners and
at greater distances from the player. See issue #39