Implemented simple play-as-render code
This is a very rudimentary implementation of a 'play-as-render'
function. Currently allows for manual player movement with the AI
controlling everything else. Can be activated with
createRenderUser(%client);

Several instances of %render.delete(); have been replaced with a
function, including instances in the debug add-on.

Also includes fixes for console errors caused by victim bots.