Render can no-longer jet
This is due to complications with the AI (proper smart jetting is
difficult to accomplish, especially without breaking immersion with
workarounds).

It is also difficult to disable jets properly for Render players without
creating extra datablocks and switching them in real time.

This also prevents Render players from 'crouch jumping' to gain an
advantage.